# Speeding up shortstr/longstr (de)serialization.