@danielmarbach 2020-12-10
Out of curiosity why are you not using the [Benchmark Baseline](https://benchmarkdotnet.org/articles/features/baselines.html#sample-introbenchmarkbaseline) for the comparison?