@stebet 2020-12-10
> Out of curiosity why are you not using the [Benchmark Baseline](https://benchmarkdotnet.org/articles/features/baselines.html#sample-introbenchmarkbaseline) for the comparison?

I didn't want to copy the old logic over to the new one since I'm manually comparing runs between different branches so there's no straightforward way to add the baseline parameter. There is a tool that @adamsitnik has mentioned that can compare benchmarkdotnet runs but I haven't tried it yet. See here: https://github.com/dotnet/BenchmarkDotNet/issues/973

Hopefully it'll become a global tool soon to make this easier :)