@adamsitnik 2020-12-10
> Hopefully it'll become a global tool soon to make this easier

We don't have such plans for the near future, but it's available [here](https://github.com/dotnet/performance/tree/master/src/tools/ResultsComparer) if someone wants to use it