@stebet 2020-12-10
> @stebet any reason not to backport this all the way to `6.x`?

Don't think so :)